--- src/origin/hooks/useCart.ts
@@ -1,59 +1,0 @@
-// useCart.ts
-import { useState } from 'react';
-import { CartItem, Coupon, Product } from '../../types.ts';
-import {
-  calculateCartTotal,
-  updateCartItemQuantity
-} from './utils/cartUtils.ts';
-
-export const useCart = () => {
-  const [cart, setCart] = useState<CartItem[]>([]);
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
-
-  const addToCart = (product: Product) => {
-    setCart((prevCart) => {
-      const isInCart = prevCart.find((item) => item.product.id === product.id);
-
-      if (isInCart) {
-        return prevCart.map((item) =>
-          item.product.id === product.id
-            ? {
-                ...item,
-                quantity: Math.max(item.quantity + 1, product.stock)
-              }
-            : item
-        );
-      } else {
-        return [...prevCart, { product, quantity: 1 }];
-      }
-    });
-  };
-
-  const removeFromCart = (productId: string) => {
-    setCart((prevCart) =>
-      prevCart.filter((item) => item.product.id !== productId)
-    );
-  };
-
-  const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart((prevCart) =>
-      updateCartItemQuantity(prevCart, productId, newQuantity)
-    );
-  };
-
-  const applyCoupon = (coupon: Coupon) => {
-    setSelectedCoupon(coupon);
-  };
-
-  const calculateTotal = () => calculateCartTotal(cart, selectedCoupon);
-
-  return {
-    cart,
-    addToCart,
-    removeFromCart,
-    updateQuantity,
-    applyCoupon,
-    calculateTotal,
-    selectedCoupon
-  };
-};
